+++ CHANGELOG.md
@@ -70,6 +70,20 @@ We used to have incremental versioning before `0.1.0`.
 - Fixes a [crash](https://github.com/wemake-services/wemake-python-styleguide/issues/423) with class attributes assignment
 
 
+## Version 0.6.3
+
+### Bugfixes
+
+- Fixes a [crash](https://github.com/wemake-services/wemake-python-styleguide/issues/450) with `dict`s with just values and no keys
+
+
+## Version 0.6.2
+
+### Bugfixes
+
+- Fixes a [crash](https://github.com/wemake-services/wemake-python-styleguide/issues/423) with class attributes assignment
+
+
 ## Version 0.6.1
 
 ### Bugfixes